use &mut RCC

+++ src/fmpi2c.rs
@@ -198,7 +198,7 @@ fn calculate_timing(
     let mut presc: u8;
     // if ratio is > (scll+sclh)*presc. that frequancy is not possible to generate. so
     // minimum frequancy possible is generated
-    if product > 8192 as f32 {
+    if product > 8192_f32 {
         // TODO: should we panic or use minimum allowed frequancy
         scl_l = 0x7fu8;
         scl_h = 0x7fu8;
@@ -214,7 +214,7 @@ fn calculate_timing(
             let deviation = product % tmp_presc as f32;
             if min_deviation > deviation {
                 min_deviation = deviation;
-                presc = tmp_presc as u8;
+                presc = tmp_presc;
             }
         }
         // now that we have optimal prescalar value. optimal scl_l and scl_h
